+++ CHANGELOG.md
@@ -34,6 +34,13 @@ New Features
   pupil plane. (#601)
 - Added `nx`, `ny`, and `bounds` keywords to drawImage() and drawKImage()
   methods. (#603)
+- Added `InterpolatedChromaticObject` class that can facilitate faster drawing
+  compared to brute force for chromatic objects with basic properties that are
+  wavelength-dependent (e.g., optical PSFs).  However, it can also be used to
+  carry out the brute force comparison for easy accuracy tests.  New
+  `ChromaticOpticalPSF` method takes advantage of the
+  `InterpolatedChromaticObject` class, allowing the diffraction limit and
+  aberrations to be wavelength-dependent. (#618)
 
 Bug Fixes and Improvements
 --------------------------